[Network] Remove legacy Windows Winsock message-pump backend and finalise IOCP as the permanent Windows network backend
* Removed net_windows.cpp, win32/winsockwrappers.cpp and win32/winsockwrappers.h (superseded by IOCP)
* Dropped ENABLE_IOCP CMake option; IOCP is now unconditionally active on Windows
* [Core] Removed RES::EXCEPTION_HANDLER and RES::NET_PROCESSING resource types and the NETMSG enum
* [Core] Removed glNetProcessMessages hook calls from the message loop
* Added Windows registry proxy detection to net_iocp.cpp
